Function ever_client::abi::encode_internal_message

pub async fn encode_internal_message(
    context: Arc<ClientContext>, 
    params: ParamsOfEncodeInternalMessage
) -> ClientResult<ResultOfEncodeInternalMessage>

Encodes an internal ABI-compatible message

Allows to encode deploy and function call messages.

Use cases include messages of any possible type:

deploy with initial function call (i.e. constructor or any other function that is used for some kind of initialization);

deploy without initial function call;

simple function call

There is an optional public key can be provided in deploy set in order to substitute one in TVM file.

Public key resolving priority:

Public key from deploy set.

Public key, specified in TVM file.