Module ton_client::boc

Structs

BocConfig

Error

ParamsOfBocCacheGet

ParamsOfBocCacheSet

ParamsOfBocCacheUnpin

ParamsOfEncodeBoc

ParamsOfGetBlockchainConfig

ParamsOfGetBocHash

ParamsOfGetCodeFromTvc

ParamsOfParse

ParamsOfParseShardstate

ResultOfBocCacheGet

ResultOfBocCacheSet

ResultOfEncodeBoc

ResultOfGetBlockchainConfig

ResultOfGetBocHash

ResultOfGetCodeFromTvc

ResultOfParse

Enums

BocCacheType

BuilderOp
Cell builder operation.

ErrorCode

Functions

cache_get
Get BOC from cache

cache_set
Save BOC into cache

cache_unpin
Unpin BOCs with specified pin. BOCs which don’t have another pins will be removed from cache

default_cache_max_size

encode_boc
Encodes BOC from builder operations.

get_blockchain_config
Extract blockchain configuration from key block and also from zerostate.

get_boc_hash
Calculates BOC root hash

get_code_from_tvc
Extracts code from TVC contract image

parse_account
Parses account boc into a JSON

parse_block
Parses block boc into a JSON

parse_message
Parses message boc into a JSON

parse_shardstate
Parses shardstate boc into a JSON

parse_transaction
Parses transaction boc into a JSON

required_boc

source_boc