Function ton_client::crypto::nacl_sign_keypair_from_secret_key

pub fn nacl_sign_keypair_from_secret_key(
    _context: Arc<ClientContext>, 
    params: ParamsOfNaclSignKeyPairFromSecret
) -> ClientResult<KeyPair>

Generates a key pair for signing from the secret key

NOTE: In the result the secret key is actually the concatenation of secret and public keys (128 symbols hex string) by design of NaCL. See also the stackexchange question.