Module ever_client::debot

Structs

DAction
Describes a debot action in a Debot Context.

DContext
Debot Context. Consists of several actions.

DEngine
Debot Engine. Downloads and stores debot, executes its actions and calls Debot Browser callbacks.

DebotAction
UNSTABLE Describes a debot action in a Debot Context.

DebotHandle
UNSTABLE Handle of registered in SDK debot

DebotInfo
UNSTABLE Describes DeBot metadata.

Error

ParamsOfExecute
UNSTABLE Parameters for executing debot action.

ParamsOfFetch
UNSTABLE Parameters to fetch DeBot metadata.

ParamsOfInit
UNSTABLE Parameters to init DeBot.

ParamsOfRemove
UNSTABLE

ParamsOfSend
UNSTABLE Parameters of send function.

ParamsOfStart
UNSTABLE Parameters to start DeBot. DeBot must be already initialized with init() function.

RegisteredDebot
UNSTABLE Structure for storing debot handle returned from init function.

ResultOfFetch
UNSTABLE

Spending
UNSTABLE Describes how much funds will be debited from the target contract balance as a result of the transaction.

Enums

DebotActivity
UNSTABLE Describes the operation that the DeBot wants to perform.

ErrorCode

Constants

DEBOT_WC

STATE_EXIT

STATE_ZERO

Traits

BrowserCallbacks
Callbacks that are called by debot engine to communicate with Debot Browser.

DebotInterface

DebotInterfaceExecutor

Functions

execute
UNSTABLE Executes debot action.

execute_api

fetch
UNSTABLE Fetches DeBot metadata from blockchain.

fetch_api

init
UNSTABLE Creates an instance of DeBot.

remove
UNSTABLE Destroys debot handle.

remove_api

send
UNSTABLE Sends message to Debot.

send_api

start
UNSTABLE Starts the DeBot.

start_api

Type Definitions

InterfaceResult