Trait ton_client::debot::BrowserCallbacks

pub trait BrowserCallbacks {
#[must_use]    fn log<'life0, 'async_trait>(
        &'life0 self, 
        msg: String
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn switch<'life0, 'async_trait>(
        &'life0 self, 
        ctx_id: u8
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn switch_completed<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn show_action<'life0, 'async_trait>(
        &'life0 self, 
        act: DAction
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn input<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        prompt: &'life1 str, 
        value: &'life2 mut String
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_signing_box<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<SigningBoxHandle, String>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn invoke_debot<'life0, 'async_trait>(
        &'life0 self, 
        debot: String, 
        action: DAction
    ) -> Pin<Box<dyn Future<Output = Result<(), String>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn send<'life0, 'async_trait>(
        &'life0 self, 
        message: String
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn approve<'life0, 'async_trait>(
        &'life0 self, 
        activity: DebotActivity
    ) -> Pin<Box<dyn Future<Output = ClientResult<bool>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Callbacks that are called by debot engine to communicate with Debot Browser.

Required methods

#[must_use]fn log<'life0, 'async_trait>(
&'life0 self,
msg: String
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Prints text message to user.

#[must_use]fn switch<'life0, 'async_trait>(
&'life0 self,
ctx_id: u8
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Notify that debot is switched to another context.

#[must_use]fn switch_completed<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Notify that all actions are shown to user and switch to conetxt is completed.

#[must_use]fn show_action<'life0, 'async_trait>(
&'life0 self,
act: DAction
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Show action to the user as menu item. Called after switch callback for every action in context.

#[must_use]fn input<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
prompt: &'life1 str,
value: &'life2 mut String
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Requests input from user.

#[must_use]fn get_signing_box<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<SigningBoxHandle, String>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Requests keys from user.

#[must_use]fn invoke_debot<'life0, 'async_trait>(
&'life0 self,
debot: String,
action: DAction
) -> Pin<Box<dyn Future<Output = Result<(), String>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Executes action of another debot.

#[must_use]fn send<'life0, 'async_trait>(
&'life0 self,
message: String
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Sends message with debot interface call to Browser. Message parameter is a BoC encoded as Base64.

#[must_use]fn approve<'life0, 'async_trait>(
&'life0 self,
activity: DebotActivity
) -> Pin<Box<dyn Future<Output = ClientResult<bool>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Requests permission to execute DeBot operation (e.g. sending messages to blockchain).

Loading content...