Module ton_client::net

Structs

EndpointsSet

Error

FieldAggregation

MessageNode

NetworkConfig

OrderBy

ParamsOfAggregateCollection

ParamsOfBatchQuery

ParamsOfCreateBlockIterator

ParamsOfCreateTransactionIterator

ParamsOfFindLastShardBlock

ParamsOfIteratorNext

ParamsOfQuery

ParamsOfQueryCollection

ParamsOfQueryCounterparties

ParamsOfQueryTransactionTree

ParamsOfResumeBlockIterator

ParamsOfResumeTransactionIterator

ParamsOfSubscribeCollection

ParamsOfWaitForCollection

PostRequest

RegisteredIterator

ResultOfAggregateCollection

ResultOfBatchQuery

ResultOfFindLastShardBlock

ResultOfGetEndpoints

ResultOfIteratorNext

ResultOfQuery

ResultOfQueryCollection

ResultOfQueryTransactionTree

ResultOfSubscribeCollection

ResultOfSubscription

ResultOfWaitForCollection

TransactionNode

Enums

AggregationFn

ErrorCode

GraphQLQueryEvent

ParamsOfQueryOperation

SortDirection

SubscriptionResponseType

Constants

ACCOUNTS_COLLECTION

BLOCKS_COLLECTION

MESSAGES_COLLECTION

TRANSACTIONS_COLLECTION

Traits

ChainIterator

Functions

aggregate_collection
Aggregates collection data.

batch_query
Performs multiple queries per single fetch.

create_block_iterator
Creates block iterator.

create_transaction_iterator
Creates transaction iterator.

fetch_endpoints
Requests the list of alternative endpoints from server

fetch_endpoints_api

find_last_shard_block
Returns ID of the last block in a specified account shard

find_last_shard_block_api

get_endpoints
Requests the list of alternative endpoints from server

get_endpoints_api

iterator_next
Returns next available items.

query
Performs DAppServer GraphQL query.

query_collection
Queries collection data

query_counterparties
Allows to query and paginate through the list of accounts that the specified account has interacted with, sorted by the time of the last internal message between accounts

query_transaction_tree
Returns transactions tree for specific message.

remove_iterator
Removes an iterator

resume
Resumes network module to enable network activity

resume_api

resume_block_iterator
Resumes block iterator.

resume_transaction_iterator
Resumes transaction iterator.

set_endpoints
Sets the list of endpoints to use on reinit

set_endpoints_api

subscribe_collection

suspend
Suspends network module to stop any network activity

suspend_api

unsubscribe
Cancels a subscription

wait_for_collection
Returns an object that fulfills the conditions or waits for its appearance