Function ton_client::net::iterator_next

pub async fn iterator_next(
    context: Arc<ClientContext>, 
    params: ParamsOfIteratorNext
) -> ClientResult<ResultOfIteratorNext>

Returns next available items.

In addition to available items this function returns the has_more flag indicating that the iterator isn’t reach the end of the iterated range yet.

This function can return the empty list of available items but indicates that there are more items is available. This situation appears when the iterator doesn’t reach iterated range but database doesn’t contains available items yet.

If application requests resume state in return_resume_state parameter then this function returns resume_state that can be used later to resume the iteration from the position after returned items.

The structure of the items returned depends on the iterator used. See the description to the appropriated iterator creation function.