Enum ever_client::processing::ProcessingEvent

pub enum ProcessingEvent {
    WillFetchFirstBlock {},
    FetchFirstBlockFailed {
        error: ClientError,
    },
    WillSend {
        shard_block_id: String,
        message_id: String,
        message: String,
    },
    DidSend {
        shard_block_id: String,
        message_id: String,
        message: String,
    },
    SendFailed {
        shard_block_id: String,
        message_id: String,
        message: String,
        error: ClientError,
    },
    WillFetchNextBlock {
        shard_block_id: String,
        message_id: String,
        message: String,
    },
    FetchNextBlockFailed {
        shard_block_id: String,
        message_id: String,
        message: String,
        error: ClientError,
    },
    MessageExpired {
        message_id: String,
        message: String,
        error: ClientError,
    },
}

Variants

WillFetchFirstBlock

Notifies the app that the current shard block will be fetched from the network.

Fetched block will be used later in waiting phase.

Fields of WillFetchFirstBlock

FetchFirstBlockFailed

Notifies the app that the client has failed to fetch current shard block.

Message processing has finished.

Fields of FetchFirstBlockFailed

error: ClientError

WillSend

Notifies the app that the message will be sent to the network.

Fields of WillSend

shard_block_id: Stringmessage_id: Stringmessage: String

DidSend

Notifies the app that the message was sent to the network.

Fields of DidSend

shard_block_id: Stringmessage_id: Stringmessage: String

SendFailed

Notifies the app that the sending operation was failed with network error.

Nevertheless the processing will be continued at the waiting phase because the message possibly has been delivered to the node.

Fields of SendFailed

shard_block_id: Stringmessage_id: Stringmessage: Stringerror: ClientError

WillFetchNextBlock

Notifies the app that the next shard block will be fetched from the network.

Event can occurs more than one time due to block walking procedure.

Fields of WillFetchNextBlock

shard_block_id: Stringmessage_id: Stringmessage: String

FetchNextBlockFailed

Notifies the app that the next block can’t be fetched due to error.

Processing will be continued after network_resume_timeout.

Fields of FetchNextBlockFailed

shard_block_id: Stringmessage_id: Stringmessage: Stringerror: ClientError

MessageExpired

Notifies the app that the message was expired.

Event occurs for contracts which ABI includes header “expire”

Processing will be continued from encoding phase after expiration_retries_timeout.

Fields of MessageExpired

message_id: Stringmessage: Stringerror: ClientError